pyproject.toml
[build-system]
requires = ['setuptools>=61', 'wheel']
build-backend = 'setuptools.build_meta'

[project]
name = 'deepqmc'
version = '1.2.0'
description = 'Deep-learning quantum Monte Carlo for electrons in real space'
authors = [
    {name = "Jan Hermann"},
    {name = "Zeno Schätzle", email = "zeno.schaetzle@fu-berlin.de"},
    {name = "Péter Bernát Szabó", email = "peter.bernat.szabo@fu-berlin.de"},
]
maintainers = [
    {name = "Zeno Schätzle", email = "zeno.schaetzle@fu-berlin.de"},
    {name = "Péter Bernát Szabó", email = "peter.bernat.szabo@fu-berlin.de"},
]
readme = 'README.md'
license = {text = 'MIT'}
classifiers = [
    'Development Status :: 4 - Beta',
    'Environment :: Console',
    'Environment :: GPU :: NVIDIA CUDA',
    'Intended Audience :: Science/Research',
    'Operating System :: MacOS :: MacOS X',
    'Operating System :: POSIX :: Linux',
    'Topic :: Scientific/Engineering :: Artificial Intelligence',
    'Topic :: Scientific/Engineering :: Chemistry',
    'Topic :: Scientific/Engineering :: Physics',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
]
requires-python = ">=3.9"
dependencies = [
    'dm-haiku>0.0.9',
    'h5py',
    'hydra-core',
    'jax',
    'jaxlib',
    'jax-dataclasses',
    'kfac-jax',
    'scipy',
    'optax',
    'pyscf',
    'tensorboardX',
    'pyyaml',
    'tqdm',
    'uncertainties',
]

[project.optional-dependencies]
dev = [
    'pytest',
    'pytest-regressions',
    'black',
    'codespell',
    'flake8',
    'flake8-bugbear',
    'flake8-comprehensions',
    'flake8-quotes',
    'isort',
    'mypy',
    'pep8-naming',
    'pre-commit',
    'pydocstyle',
    'pyright',
]
doc = [
    'sphinx',
    'Jinja2',
    'sphinxcontrib-katex',
    'toml',
    'docutils',
    'pydata-sphinx-theme',
]

[project.urls]
repository = 'https://github.com/deepqmc/deepqmc'
documentation = 'https://deepqmc.github.io'

[project.scripts]
deepqmc = 'deepqmc.app:cli'

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.yaml"]

[tool.black]
target-version = ['py310']
skip-string-normalization = true
preview = true

[tool.pyright]
exclude = ['tests']
reportMissingImports = false
